# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAER301 API Documentation",
        "description": "Auto-generated API documentation for SAE301 project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/races": {
            "get": {
                "tags": [
                    "Races"
                ],
                "summary": "Get list of races",
                "description": "Returns all races with related data",
                "operationId": "bfdc03dc2b3018ad3edf09cd8632ba2b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by race name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Races retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Race"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/races/{id}": {
            "get": {
                "tags": [
                    "Races"
                ],
                "summary": "Get race by ID",
                "description": "Returns a single race with details",
                "operationId": "d40755471be0f0df220253254e69896c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Race retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Race"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Race not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Race not found"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Race": {
                "title": "Race",
                "description": "Orienteering race model",
                "properties": {
                    "race_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "race_name": {
                        "type": "string",
                        "example": "La Boussole de la Forêt"
                    },
                    "race_date_start": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-12T09:00:00Z"
                    },
                    "race_date_end": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-12T17:00:00Z"
                    },
                    "race_reduction": {
                        "type": "number",
                        "format": "float",
                        "example": 5,
                        "nullable": true
                    },
                    "race_meal_price": {
                        "type": "number",
                        "format": "float",
                        "example": 8,
                        "nullable": true
                    },
                    "race_duration_minutes": {
                        "type": "number",
                        "format": "float",
                        "example": 150,
                        "nullable": true
                    },
                    "raid_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "adh_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pac_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pae_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "dif_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "typ_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}